v1.3.12
Attempted multiframe AND animateDiff controlnet vid2vid with no success. API cannot handle such requests current. Would need to edit the multiframe script or animatediff_cn.py file to achieve this which I am lacking in expertise to do so.
Implemented depth + temporalnet which is good for the time being but could be better overall. Need to play with additional controlnet modules