v1.4.4
Tuned piano. Quotes now directly correlate to the series the quote is from. Maybe more to the specific characters down the line.
Added functionality to check if the stable diffusion server is on, and if not will start it. Need to implement a kill switch next so the server doesn't remain on if it does not need to be